16-OnlineShop/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic==2.12.5
pydantic-settings==2.7.0
email-validator==2.3.0
PyJWT==2.11.0
bcrypt==5.0.0
stripe==14.4.0
python-multipart==0.0.22
jinja2==3.1.2
aiofiles==25.1.0
